move useStreamingUIMessages back out
